+++ pkg/tests/cross-tests/input_check.go
@@ -19,10 +19,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/pulumi/providertest/providers"
-	"github.com/pulumi/providertest/pulumitest"
-	"github.com/pulumi/providertest/pulumitest/opttest"
-	shimv2 "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfshim/sdk-v2"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tests/pulcheck"
 	"github.com/stretchr/testify/require"
 )
 
@@ -120,7 +117,6 @@ func runCreateInputCheck(t T, tc inputTestCase) {
 		rtype             = "crossprovider_testres"
 		rtok              = "TestRes"
 		rtoken            = providerShortName + ":index:" + rtok
-		providerVer       = "0.0.1"
 	)
 
 	tfwd := t.TempDir()
@@ -128,42 +124,17 @@ func runCreateInputCheck(t T, tc inputTestCase) {
 	tfd := newTfDriver(t, tfwd, providerShortName, rtype, tc.Resource)
 	tfd.writePlanApply(t, tc.Resource.Schema, rtype, "example", tc.Config)
 
-	tfp := &schema.Provider{
-		ResourcesMap: map[string]*schema.Resource{
-			rtype: tc.Resource,
-		},
-	}
-	ensureProviderValid(t, tfp)
-
-	shimProvider := shimv2.NewProvider(tfp, shimv2.WithPlanResourceChange(
-		func(tfResourceType string) bool { return true },
-	))
-
+	resMap := map[string]*schema.Resource{rtype: tc.Resource}
+	bridgedProvider := pulcheck.BridgedProvider(t, providerShortName, resMap)
 	pd := &pulumiDriver{
 		name:                providerShortName,
-		version:             providerVer,
-		shimProvider:        shimProvider,
 		pulumiResourceToken: rtoken,
 		tfResourceName:      rtype,
 		objectType:          tc.ObjectType,
 	}
+	yamlProgram := pd.generateYAML(t, bridgedProvider.P.ResourcesMap() ,tc.Config)
 
-	puwd := t.TempDir()
-	pd.writeYAML(t, puwd, tc.Config)
-
-	pt := pulumitest.NewPulumiTest(t, puwd,
-		opttest.TestInPlace(),
-		opttest.SkipInstall(),
-		opttest.AttachProvider(
-			providerShortName,
-			func(ctx context.Context, pt providers.PulumiTest) (providers.Port, error) {
-				handle, err := pd.startPulumiProvider(ctx)
-				require.NoError(t, err)
-				return providers.Port(handle.Port), nil
-			},
-		),
-		opttest.Env("DISABLE_AUTOMATIC_PLUGIN_ACQUISITION", "true"),
-	)
+	pt := pulcheck.PulCheck(t, bridgedProvider, string(yamlProgram))
 
 	pt.Up()
