+++ pkg/tests/pulcheck/pulcheck.go
@@ -1,0 +1,147 @@
+package pulcheck
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/pulumi/providertest/providers"
+	"github.com/pulumi/providertest/pulumitest"
+	"github.com/pulumi/providertest/pulumitest/opttest"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/info"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge/tokens"
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfgen"
+	shimv2 "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfshim/sdk-v2"
+	pulumidiag "github.com/pulumi/pulumi/sdk/v3/go/common/diag"
+	"github.com/pulumi/pulumi/sdk/v3/go/common/diag/colors"
+	"github.com/pulumi/pulumi/sdk/v3/go/common/util/rpcutil"
+	pulumirpc "github.com/pulumi/pulumi/sdk/v3/proto/go"
+	"github.com/stretchr/testify/require"
+	"google.golang.org/grpc"
+	"gotest.tools/assert"
+)
+
+func CheckProviderValid(tfp *schema.Provider) error {
+	for _, r := range tfp.ResourcesMap {
+		//nolint:staticcheck
+		if r.Read == nil && r.ReadContext == nil {
+			r.ReadContext = func(_ context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
+				return nil
+			}
+		}
+		if r.DeleteContext == nil {
+			r.DeleteContext = func(
+				ctx context.Context, rd *schema.ResourceData, i interface{},
+			) diag.Diagnostics {
+				return diag.Diagnostics{}
+			}
+		}
+
+		if r.CreateContext == nil {
+			r.CreateContext = func(
+				ctx context.Context, rd *schema.ResourceData, i interface{},
+			) diag.Diagnostics {
+				rd.SetId("newid")
+				return diag.Diagnostics{}
+			}
+		}
+
+		r.UpdateContext = func(
+			ctx context.Context, rd *schema.ResourceData, i interface{},
+		) diag.Diagnostics {
+			return diag.Diagnostics{}
+		}
+	}
+	return tfp.InternalValidate()
+}
+
+func StartPulumiProvider(ctx context.Context, name, version string, providerInfo tfbridge.ProviderInfo) (*rpcutil.ServeHandle, error) {
+	sink := pulumidiag.DefaultSink(io.Discard, io.Discard, pulumidiag.FormatOptions{
+		Color: colors.Never,
+	})
+
+	schema, err := tfgen.GenerateSchema(providerInfo, sink)
+	if err != nil {
+		return nil, fmt.Errorf("tfgen.GenerateSchema failed: %w", err)
+	}
+
+	schemaBytes, err := json.MarshalIndent(schema, "", " ")
+	if err != nil {
+		return nil, fmt.Errorf("json.MarshalIndent(schema, ..) failed: %w", err)
+	}
+
+	prov := tfbridge.NewProvider(ctx, nil, name, version, providerInfo.P, providerInfo, schemaBytes)
+
+	handle, err := rpcutil.ServeWithOptions(rpcutil.ServeOptions{
+		Init: func(srv *grpc.Server) error {
+			pulumirpc.RegisterResourceProviderServer(srv, prov)
+			return nil
+		},
+	})
+	if err != nil {
+		return nil, fmt.Errorf("rpcutil.ServeWithOptions failed: %w", err)
+	}
+
+	return &handle, nil
+}
+
+type T interface {
+	Logf(string, ...any)
+	TempDir() string
+	require.TestingT
+	assert.TestingT
+	pulumitest.PT
+}
+
+func BridgedProvider(t T, providerName string, resMap map[string]*schema.Resource) info.Provider {
+	tfp := &schema.Provider{ResourcesMap: resMap}
+	err := CheckProviderValid(tfp)
+	require.NoError(t, err)
+
+	shimProvider := shimv2.NewProvider(tfp, shimv2.WithPlanResourceChange(
+		func(tfResourceType string) bool { return true },
+	))
+
+	provider := tfbridge.ProviderInfo{
+		P:            shimProvider,
+		Name:         providerName,
+		Version:      "0.0.1",
+		MetadataInfo: &tfbridge.MetadataInfo{},
+	}
+	makeToken := func(module, name string) (string, error) {
+		return tokens.MakeStandard(providerName)(module, name)
+	}
+	provider.MustComputeTokens(tokens.SingleModule("prov", "index", makeToken))
+
+	return provider
+}
+
+func PulCheck(t T, bridgedProvider info.Provider, program string) *pulumitest.PulumiTest {
+	puwd := t.TempDir()
+	p := filepath.Join(puwd, "Pulumi.yaml")
+
+	err := os.WriteFile(p, []byte(program), 0o600)
+	require.NoError(t, err)
+
+	opts := []opttest.Option{
+		opttest.Env("DISABLE_AUTOMATIC_PLUGIN_ACQUISITION", "true"),
+		opttest.TestInPlace(),
+		opttest.SkipInstall(),
+		opttest.AttachProvider(
+			bridgedProvider.Name,
+			func(ctx context.Context, pt providers.PulumiTest) (providers.Port, error) {
+				handle, err := StartPulumiProvider(ctx, bridgedProvider.Name, bridgedProvider.Version, bridgedProvider)
+				require.NoError(t, err)
+				return providers.Port(handle.Port), nil
+			},
+		),
+	}
+
+	return pulumitest.NewPulumiTest(t, puwd, opts...)
+}
